python_yolo/requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pillow
numpy
opencv-python-headless
# For GPU support (optional but good to have)
# torch --index-url https://download.pytorch.org/whl/cu118
